+++ mapper/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.1
 
 * #131, `@JsonProperty` annotation in superclass should be inherited for overriden getters
+* `deserialize` method extended to accept json `String` or parsed json `Object`
 
 ## 2.1.0